Refactor Home Assistant video camera modules
